feat: add alert prompt
