+++ src/components/Cart/ConfirmationDialog/ConfirmationDialog.tsx
@@ -1,0 +1,46 @@
+// import * as React from 'react';
+// import Button from '@mui/material/Button';
+// import Dialog from '@mui/material/Dialog';
+// import DialogActions from '@mui/material/DialogActions';
+// import DialogContent from '@mui/material/DialogContent';
+// import DialogContentText from '@mui/material/DialogContentText';
+// import DialogTitle from '@mui/material/DialogTitle';
+
+// export default function AlertDialog() {
+//   const [open, setOpen] = React.useState(false);
+
+//   // const handleClickOpen = () => {
+//   //   setOpen(true);
+//   // };
+
+//   const handleClose = () => {
+//     setOpen(false);
+//   };
+
+//   return (
+//     <div>
+//       <Dialog
+//         open={open}
+//         onClose={handleClose}
+//         aria-labelledby="alert-dialog-title"
+//         aria-describedby="alert-dialog-description"
+//       >
+//         <DialogTitle id="alert-dialog-title">
+//           {"Use Google's location service?"}
+//         </DialogTitle>
+//         <DialogContent>
+//           <DialogContentText id="alert-dialog-description">
+//             Let Google help apps determine location. This means sending anonymous
+//             location data to Google, even when no apps are running.
+//           </DialogContentText>
+//         </DialogContent>
+//         <DialogActions>
+//           <Button onClick={handleClose}>Disagree</Button>
+//           <Button onClick={handleClose} autoFocus>
+//             Agree
+//           </Button>
+//         </DialogActions>
+//       </Dialog>
+//     </div>
+//   );
+// }
